working version without saved accounts

+++ src/user/PersonCreator.java
@@ -14,15 +14,15 @@ public class PersonCreator {
         System.out.println("Jak masz na nazwisko:"); // add option: give last name or skip it
         String lastName = dataReader.getString();
         System.out.println("Jaką masz stawkę godzinową (netto):");
-        double wage = dataReader.getDouble();
+        int wage = dataReader.getInt();
         System.out.println("Ile płacisz barmanowi:"); //add option YES or NOT
-        double bartender = dataReader.getDouble(); // add percentage option
+        int bartender = dataReader.getInt(); // add percentage option
         System.out.println("Srednio ile dziennie wydajesz na posiłki:");
-        double dinner = dataReader.getDouble();
+        int dinner = dataReader.getInt();
         System.out.println("Ile płacisz na zmywak:");
-        double sink = dataReader.getDouble();
+        int sink = dataReader.getInt();
         System.out.println("Dodatkowe dzienne koszty:");
-        double other = dataReader.getDouble();
+        int other = dataReader.getInt();
 
         return new Person(nickname, firstName, lastName, wage, bartender, dinner, sink, other);
     }
